requirements.txt
# Core dependencies
requests>=2.31.0
pandas>=2.1.0
geopandas>=0.14.0
PyYAML>=6.0.1
shapely>=2.0.0
tqdm>=4.66.0

# Sentinel-1 specific
stsa>=0.1.0  # Sentinel-1 Tops Split Analyzer

# Sentinel-2 specific
earthengine-api>=0.1.400  # Google Earth Engine API

# Optional but recommended
numpy>=1.24.0  # Required by pandas and geopandas
rasterio>=1.3.0  # For raster operations
pyproj>=3.6.0  # For coordinate transformations

# Development dependencies
pytest>=7.4.0  # For testing
black>=23.7.0  # For code formatting
flake8>=6.1.0  # For linting 